company page save button implemented..

+++ src/accounts/app/dao/UserDao.java
@@ -1,5 +1,7 @@
 package accounts.app.dao;
 
+import org.springframework.stereotype.Repository;
+
 import accounts.model.entity.user.User;
 
 /**
@@ -7,6 +9,7 @@ import accounts.model.entity.user.User;
  * @author syed
  *
  */
+
 public interface UserDao extends GenericDao<User>{
-    User getUser(int userId);
+    User getUser(long userId);
 }
